requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
typer>=0.9.0
rich>=10.0.0
tqdm>=4.65.0
PyQt6>=6.4.0
PyQt6-WebEngine>=6.4.0
PyQt6-Qt6>=6.4.0
PyQt6-sip>=13.4.0
pyinstaller>=5.0.0
Pillow>=9.0.0
numpy>=1.20.0
torch>=2.0.0
torchaudio>=2.0.0
whisper>=1.1.10 